[SESSION] Implement basic version of wordle

+++ SESSION/src/main/java/org/example/session/service/WordleService.java
@@ -8,5 +8,5 @@ public interface WordleService {
 
 	Mono<Session> initializeSession(int userId, int languageId, int wordLength);
 
-	Mono<WordleResult> handleGuess(int userId, int sessionId, String guess);
+	Mono<WordleResult> handleGuess(int userId, long sessionId, String guess);
 }
